requirements.txt
# FastAPI and ASGI Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Google Gemini AI
google-generativeai==0.3.2

# XML/XAML Parsing
lxml==5.1.0
xmltodict==0.13.0

# Utilities
aiofiles==23.2.1
python-dateutil==2.8.2

# CORS
fastapi-cors==0.0.6

# Optional: Vector DB for RAG (if you want advanced features)
# chromadb==0.4.22
# langchain==0.1.0
# langchain-google-genai==0.0.6
